chore: full route url for all pages
